Mention ReferenceStorageTypes in diagnostics
Occasionally, you’ll get a paradoxical error from the compiler like this one:

```
warning: property ‘foo’ of type ‘Foo?’ does not match type ‘Foo?’ declared by the header
```

One of the ways this can happen is when a reference ownership modifier, like `weak` or `unowned(unsafe)`, is applied to the declaration. This causes an extra ReferenceStorageType to be wrapped around the type, but this type information was previously not printed in diagnostics; the diagnostic would mention two seemingly-identical types even though the compiler can tell the difference between them and this difference may actually be why the error occurred.

Modify PrintOptions to include an option that prints ReferenceStorageType in a developer-readable form, and use this option when DiagnosticEngine prints a Type. This won’t matter for most diagnostics—the constraint solver, for instance, systematically strips ReferenceStorageTypes out of the types it works with—but when this information *is* present, it’ll now be visible in the error message. That turns the above diagnostic into:

```
warning: property ‘foo’ of type ‘Foo?’ does not match type ‘unowned(unsafe) Foo?’ declared by the header
```

The reproducer I have for this (and have added a test for) involves objcImpl, but a Swift developer I spoke to mentioned they’ve seen this happen from other parts of the compiler, so I’ve chosen to make a compiler-wide change instead of tailoring the specific diagnostic.

Fixes rdar://124383038.